requirements.txt.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Django==4.0.2
django-appconf==1.0.5
django-cryptography==1.0
django-imagekit==4.1.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
isort==5.10.1
pilkit==2.0
Pillow==9.0.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.6
python-magic==0.4.25
python-magic-bin==0.4.14
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
